+++ data/update/v0.1.6/pdo_pgsql/2013-04.sql
@@ -1,0 +1,2 @@
+-- exception layout
+INSERT INTO core_config_data (identifier, value) VALUES ('site_exception_layout', '');